# Optimize revocation registry creation