This greatly reduces the time to generate a revocation registry when issuance-by-default is enabled. It also reduces the time to generate the associated tails file.

On my system, for a registry with 10,000 entries, previous sample times are:

> Generate registry for 10000 is 2.50s
> Generate tails for 10000 is 5.07s

With these updates:

> Generate registry for 10000 is 36.37ms
> Generate tails for 10000 is 3.80s